wardrobe-app-upload/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.39
pydantic-settings==2.10.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
httpx==0.28.1
Pillow==11.2.1
